# docs/swagger.yaml
...
openapi: "3.0.0"
info:
  title: Demli API
  version: 1.0.0
  description: >
    Multi-domain consumer API — rent-a-car, travel, hotel, food,
    booking flow (FlowBox), payments, reviews, wishlists.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/users/add-admin:
    post:
      tags: [Admin]
      summary: Promote a user to admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string }
      responses:
        "200": { description: User promoted }
        "403": { description: Forbidden }
  /api/admin/flowboxes:
    get:
      tags: [Admin]
      summary: List all flowboxes (admin)
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200": { description: Flowbox list }
  /api/admin/flowboxes/{id}:
    get:
      tags: [Admin]
      summary: Get flowbox by ID (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Flowbox details }
  /api/admin/flowboxes/{id}/status:
    put:
      tags: [Admin]
      summary: Update flowbox status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string }
      responses:
        "200": { description: Status updated }
...
  /api/admin/reviews/{id}:
    delete:
      tags: [Admin]
      summary: Delete a review (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Review deleted }

  # ── Services ────────────────────────────────────────────────────────────
...
  /api/food/companies/{id}:
    get:
      tags: [Food]
      summary: Get food company details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Company details }
  /api/food/items:
    get:
      tags: [Food]
      summary: List food items with filters
      security: []
      parameters:
        - in: query
          name: companyId
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: name
          schema: { type: string }
      responses:
        "200": { description: Food item list }
  /api/food/items/{id}:
    get:
      tags: [Food]
      summary: Get food item details
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Food item details }

  # ── FlowBox ─────────────────────────────────────────────────────────────
  /api/flowboxes:
    post:
      tags: [FlowBox]
      summary: Create a new FlowBox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceType, serviceId]
              properties:
                serviceType: { type: string }
                serviceId: { type: string }
      responses:
        "201": { description: FlowBox created }
    get:
      tags: [FlowBox]
      summary: List user's FlowBoxes
      responses:
        "200": { description: FlowBox list }
  /api/flowboxes/{id}:
    get:
      tags: [FlowBox]
      summary: Get FlowBox details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: FlowBox details }
  /api/flowboxes/{id}/step:
    put:
      tags: [FlowBox]
      summary: Advance FlowBox to next step
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [screen, data]
              properties:
                screen: { type: string }
                data: { type: object }
      responses:
        "200": { description: Step advanced }
  /api/flowboxes/{id}/cancel:
    put:
      tags: [FlowBox]
      summary: Cancel a FlowBox
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: FlowBox cancelled }
...
  /api/flow-screens/{serviceType}:
    get:
      tags: [FlowBox]
      summary: Get flow screen config (admin)
      parameters:
        - in: path
          name: serviceType
          required: true
          schema: { type: string }
      responses:
        "200": { description: Screen config }
    put:
      tags: [FlowBox]
      summary: Update flow screen config (admin)
      parameters:
        - in: path
          name: serviceType
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200": { description: Config updated }

  # ── Payment ─────────────────────────────────────────────────────────────
...
  /api/reviews:
    get:
      tags: [Reviews]
      summary: List reviews
      security: []
      parameters:
        - in: query
          name: targetType
          required: true
          schema: { type: string }
        - in: query
          name: targetId
          required: true
          schema: { type: string }
      responses:
        "200": { description: Review list }
    post:
      tags: [Reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId, rating]
              properties:
                targetType: { type: string }
                targetId: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        "201": { description: Review created }

  # ── Wishlist ────────────────────────────────────────────────────────────
...